10th May

+++ routes/posts.js
@@ -4,6 +4,7 @@ const router = express.Router();
 const multer = require('multer')
 const path = require('path')
 const savedPostModel = require('../models/save-post');
+const notificationModel = require('../models/notifications');
 const userModel = require('../models/users');
 const { default: mongoose } = require('mongoose');
 const { log } = require('handlebars/runtime');
@@ -138,18 +139,96 @@ router.put('/savePost', async function (req, res) {
             //     }
             // }
             // ])
+
+            // postOwner UserId
             let postUserId = await postModel.findOne({ _id: postId }, { _id: 0, userId: 1 })
+
+            // who liked the post
             let likedPostUserDetail = await userModel.findOne({ _id: userId }, { firstName: 1, lastName: 1 }).lean()
+
+            let fullName = `${likedPostUserDetail.firstName}  ${likedPostUserDetail.lastName}`
+
+            // Store Notification 
+            await notificationModel.create({ postOwnerId: postUserId.userId, likedBy: likedPostUserDetail._id, postId: postId, likedUserName: fullName })
 
             io.to(postUserId.userId.toString()).emit("newNotification", `${likedPostUserDetail.firstName} ${likedPostUserDetail.lastName} liked your post`)
 
             await savedPostModel.create({ postId: postId, userId: userId })
-            return res.send({ type: "success" })
+            res.send({ type: "success" })
+
+            let savedPostCountArray = await postModel.aggregate([
+                {
+                    $match: {
+                        _id: new mongoose.Types.ObjectId(postId)
+                    }
+                },
+                {
+                    $lookup: {
+                        from: 'savedPost',
+                        localField: '_id',
+                        foreignField: 'postId',
+                        as: 'savedPost'
+                    }
+                },
+                {
+                    $project: {
+                        'savedPostCount': { $size: '$savedPost' }
+                    }
+                }
+            ])
+
+            let savepostCount = 0
+            let savedPostObject = {}
+            for (const savedpost of savedPostCountArray) {
+                savepostCount = savedpost.savedPostCount
+                savedPostObject[postId] = savepostCount
+            }
+            io.emit("savedPostCount", JSON.stringify(savedPostObject))
         }
+
         if (req.body.title == "Unsave") {
             const postId = req.body.postId
+            const userId = req.user._id
+
             await savedPostModel.deleteOne({ postId: new mongoose.Types.ObjectId(postId) })
-            return res.send({ type: "Delete success" })
+            let postUserId = await postModel.findOne({ _id: postId }, { _id: 0, userId: 1 })
+            let likedPostUserDetail = await userModel.findOne({ _id: userId }, { firstName: 1, lastName: 1 }).lean()
+
+            //delete permanently notification if user unlike Post
+            await notificationModel.deleteOne({ postOwnerId: postUserId.userId, likedBy: likedPostUserDetail._id, postId: postId })
+
+            res.send({ type: "Delete success" })
+
+            let savedPostCountArray = await postModel.aggregate([
+                {
+                    $match: {
+                        _id: new mongoose.Types.ObjectId(postId)
+                    }
+                },
+                {
+                    $lookup: {
+                        from: 'savedPost',
+                        localField: '_id',
+                        foreignField: 'postId',
+                        as: 'savedPost'
+                    }
+                },
+                {
+                    $project: {
+                        'savedPostCount': { $size: '$savedPost' }
+                    }
+                }
+            ])
+
+            let savepostCount = 0
+            let savedPostObject = {}
+            for (const savedpost of savedPostCountArray) {
+                savepostCount = savedpost.savedPostCount
+                savedPostObject[postId] = savepostCount
+            }
+
+
+            io.emit("savedPostCount", JSON.stringify(savedPostObject))
         }
 
     } catch (error) {
@@ -184,7 +263,6 @@ router.get('/', async function (req, res) {
 })
 
 router.put('/postId', async function (req, res, next) {
-
     try {
         let maxSize = 2000000;
         const upload = multer({
@@ -199,33 +277,62 @@ router.put('/postId', async function (req, res, next) {
             let { editTitle, editDescription, postPath } = req.body
             let findPostpath = await postModel.findOne({ "_id": req.body.editPostId }, { 'postPath': 1 }) // find old postPath by postId
             let oldPostPath = findPostpath.postPath
-            console.log("=========", oldPostPath);
 
-            fs.unlink('public' + `/images/posts/${oldPostPath}`, (err => {
-                if (err) console.log(err);
-                else {
-                    console.log("11111111");
+            if (req.file?.filename) {
+                fs.unlink('public' + `/images/posts/${oldPostPath}`, (err => {
+                    if (err) console.log(err);
+                    else {
+                        console.log("unlinked path");
+                    }
+                }));
+                let updateQuery = { "title": editTitle, "description": editDescription, 'postPath': postPath }
+                if (req.fileValidationError) {
+                    return res.send({ type: 'error' });
                 }
-            }));
-            let updateQuery = { "title": editTitle, "description": editDescription, 'postPath': postPath }
-            if (req.fileValidationError) {
-                return res.send({ type: 'error' });
-            }
-            if (err) {
-                return res.send({ type: 'tooLarge' });
-            }
-            else if (req.file) {
-                updateQuery['postPath'] = req.file?.filename
-            }
-            req.body.postPath = req.file?.filename
+                if (err) {
+                    return res.send({ type: 'tooLarge' });
+                }
+                else if (req.file) {
+                    updateQuery['postPath'] = req.file?.filename
+                }
+                req.body.postPath = req.file?.filename
 
-            await postModel.updateOne({ "_id": req.body.editPostId }, { $set: updateQuery })
-            return res.send({ type: "success" })
+                await postModel.updateOne({ "_id": req.body.editPostId }, { $set: updateQuery })
+                return res.send({ type: "success" })
+            } else {
+                return res.send({ type: "success" })
+            }
+
         });
     } catch (error) {
         console.log(error)
         return res.send({ type: "error" })
     }
 })
+
+// router.get('/notification', async function (req, res) {
+//     try {
+//         let notificationData = await notificationModel.find({ postOwnerId: req.user._id , isSeen : false }).sort({ createdOn : -1 }).lean()
+//         return res.render('partials/notification', {
+//             notificationData: notificationData
+//         })
+//     } catch (error) {
+//         console.log(error);
+//     }
+// })
+
+router.get('/get-post-data/:postId', async function (req, res) {
+    try {
+        let postData = await postModel.findOne({ _id: req.params.postId }).lean()
+        let notificationIsSeen = await notificationModel.updateOne({ postId : req.params.postId } , { $set : { isSeen : true } })
+        console.log(notificationIsSeen);
+        return res.send(postData)
+
+    } catch (error) {
+        console.log(error);
+    }
+})
+
+
 
 module.exports = router;
